+++ src/main/java/com/pullanner/api/controller/oauth2/dto/OAuth2Provider.java
@@ -1,4 +1,4 @@
-package com.pullanner.global.auth.oauth2.dto;
+package com.pullanner.api.controller.oauth2.dto;
 
 public enum OAuth2Provider {
     GOOGLE, NAVER, KAKAO, NONE;
