+++ src/main/java/com/pullanner/api/service/user/UserService.java
@@ -5,12 +5,12 @@ import static com.pullanner.api.ApiUtil.getResponseEntity;
 import com.pullanner.api.controller.user.dto.UserResponse;
 import com.pullanner.api.controller.user.dto.UserUpdateRequest;
 import com.pullanner.domain.user.User;
-import com.pullanner.domain.user.InvalidMailAuthorizationCodeException;
+import com.pullanner.exception.user.InvalidMailAuthorizationCodeException;
 import com.pullanner.domain.user.MailAuthorizationCodeRepository;
 import com.pullanner.domain.user.UserRepository;
 import com.pullanner.api.ApiResponseCode;
 import com.pullanner.api.ApiResponseMessage;
-import com.pullanner.global.auth.jwt.repository.RefreshTokenRepository;
+import com.pullanner.domain.token.RefreshTokenRepository;
 import java.util.concurrent.ThreadLocalRandom;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
