+++ src/main/java/com/pullanner/exception/token/HackedTokenException.java
@@ -1,4 +1,4 @@
-package com.pullanner.global.auth.jwt.exception;
+package com.pullanner.exception.token;
 
 public class HackedTokenException extends RuntimeException {
